refactor: move blockaid utilities & types to @safe-global/utils
